requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ruamel.yaml==0.18.10
